fix: apply stash change
