+++ payment/src/main/java/untitled/domain/PointChargeRequested.java
@@ -1,0 +1,17 @@
+package untitled.domain;
+
+import java.util.*;
+import lombok.*;
+import untitled.domain.*;
+import untitled.infra.AbstractEvent;
+
+@Data
+@ToString
+public class PointPaymentRequested extends AbstractEvent {
+
+    private Long id;
+    private Long readerId;
+    private Integer point;
+    private String impUid;
+    private Integer cost; // 결제 금액
+}
